refactor(scheduler): Extract ModuleStateTracker for SRP compliance (#1557)
Extract state tracking responsibilities from ModuleScheduler into a dedicated
ModuleStateTracker class to improve Single Responsibility Principle compliance.

Changes:
- Add IModuleStateTracker interface defining state transition operations
- Add ModuleStateTracker class implementing state management logic
- Update ModuleScheduler to delegate state operations to ModuleStateTracker
- Move MarkModuleStarted, MarkModuleCompleted, CancelPendingModules,
  GetModuleState, GetModuleCompletionTask, and NotifyDependentModules
  to the new ModuleStateTracker class

This reduces ModuleScheduler from ~780 lines to ~620 lines and separates
the state management concern from scheduling logic.
